Split too long wizard file
