+++ com.unity.render-pipelines.high-definition/Editor/DefaultScene/HDWizard.Configuration.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 14af1616d78f70d479e35be0acb47d57
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
